Allow scriptExtensionsTest option to be overridden in config
It's an edge case really, but I need to have a .js file extracted from the bundle (which has a .js extension) and the plugin will incorrectly remove that .js file and not only the bundle. So this option will allow me to set an unusual extension for the bundle (like .rem.js) and setup the plugin to look for that instead.